+++ src/softaware.Cqs.Tests/CQ.Contract/Commands/CommandWithDependency.cs
@@ -1,0 +1,6 @@
+﻿namespace softaware.Cqs.Tests.CQ.Contract.Commands
+{
+    public class CommandWithDependency : ICommand
+    {
+    }
+}