+++ src/softaware.Cqs.Tests/Fakes/Dependency.cs
@@ -1,0 +1,10 @@
+﻿namespace softaware.Cqs.Tests.Fakes
+{
+    public class Dependency : IDependency
+    {
+        public void SomeMethod()
+        {
+            // Nothing to do.
+        }
+    }
+}