+++ src/softaware.Cqs.Tests/Fakes/IDependency.cs
@@ -1,0 +1,7 @@
+﻿namespace softaware.Cqs.Tests.Fakes
+{
+    public interface IDependency
+    {
+        void SomeMethod();
+    }
+}