+++ src/softaware.Cqs.Tests/FluentValidationDecoratorTest.cs
@@ -10,6 +10,7 @@ using SimpleInjector;
 using softaware.Cqs.Decorators.FluentValidation;
 using softaware.Cqs.Tests.CQ.Contract.Commands;
 using softaware.Cqs.Tests.CQ.Contract.Queries;
+using softaware.Cqs.Tests.Fakes;
 
 namespace softaware.Cqs.Tests
 {
@@ -146,6 +147,8 @@ namespace softaware.Cqs.Tests
 
                 this.container.Collection.Register(typeof(IValidator<>), Assembly.GetExecutingAssembly());
 
+                this.container.Register<IDependency, Dependency>();
+
                 this.container.Verify();
 
                 base.SetUp();
@@ -172,7 +175,13 @@ namespace softaware.Cqs.Tests
                     .AddDecorators(b => b.AddFluentValidationDecorators(
                         builder => builder.IncludeTypesFrom(Assembly.GetExecutingAssembly())));
 
-                this.serviceProvider = services.BuildServiceProvider();
+                services.AddTransient<IDependency, Dependency>();
+
+                this.serviceProvider = services.BuildServiceProvider(new ServiceProviderOptions
+                {
+                    ValidateOnBuild = true,
+                    ValidateScopes = true
+                });
 
                 base.SetUp();
             }