+++ src/softaware.Cqs.Tests/TransactionDecoratorTest.cs
@@ -8,6 +8,7 @@ using SimpleInjector;
 using softaware.Cqs.Decorators.Transaction;
 using softaware.Cqs.Tests.CQ.Contract.Commands;
 using softaware.Cqs.Tests.CQ.Contract.Queries;
+using softaware.Cqs.Tests.Fakes;
 
 namespace softaware.Cqs.Tests
 {
@@ -112,6 +113,8 @@ namespace softaware.Cqs.Tests
                         .AddQueryHandlerDecorator(typeof(TransactionAwareQueryHandlerDecorator<,>))
                         .AddCommandHandlerDecorator(typeof(TransactionAwareCommandHandlerDecorator<>)));
 
+                this.container.Register<IDependency, Dependency>();
+
                 this.container.Verify();
 
                 base.SetUp();
@@ -137,7 +140,13 @@ namespace softaware.Cqs.Tests
                         .AddTransactionCommandHandlerDecorator()
                         .AddTransactionQueryHandlerDecorator());
 
-                this.serviceProvider = services.BuildServiceProvider();
+                services.AddTransient<IDependency, Dependency>();
+
+                this.serviceProvider = services.BuildServiceProvider(new ServiceProviderOptions
+                {
+                    ValidateOnBuild = true,
+                    ValidateScopes = true
+                });
 
                 base.SetUp();
             }